refactor(poderes): extrai o catálogo do editor para módulo compartilhado
`usePowerCatalog` e `CatalogPanel` já eram independentes do editor — recebem
dados puros e callbacks, não sabem de `CharacterSheet` nem de salvar. Só
estavam morando dentro de `PowersEditor`, o que impedia o assistente de
subida de nível de reusá-los.

Move para `components/PowerCatalog` sem mudar comportamento. Os tipos de
entrada do catálogo (`PowerCategory`, `ClassPowerSet`, `ClassAbilitySet`)
vão junto: são o contrato do catálogo, não do editor, e viviam em
`usePowersEditor` só por acidente de origem.
